requirements.txt
kafka-python
boto3
snowflake-connector-python
dbt-core
dbt-snowflake
requests


